bootcamp/kis_mock_trading/requirements.txt
# KIS API
requests>=2.31.0
python-dotenv>=1.0.1
